Configurable lifebar coloring

+++ src/include/unit.h
@@ -364,7 +364,7 @@ public:
 	unsigned Boarded : 1;        /// Unit is on board a transporter.
 	unsigned CacheLock : 1;      /// Unit is on lock by unitcache operations.
 
-	unsigned Summoned : 1;       /// Unit is summoned using spells.
+	unsigned long Summoned : 1;  /// Unit is summoned using spells.
 	unsigned Waiting : 1;        /// Unit is waiting and playing its still animation
 	unsigned MineLow : 1;        /// This mine got a notification about its resources being low
